requirements.txt
pytest>=7.0.0
hypothesis>=6.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
python-dotenv>=1.0.0
google-genai>=1.0.0
python-multipart>=0.0.6
